spm support
