+++ .gitignore
@@ -18,3 +18,7 @@ DerivedData
 *.ipa
 *.xcuserstate
 .DS_Store
+derived_data
+.swiftpm
+.*.sw[nop]
+
