+++ Demos/Demos.xcworkspace/contents.xcworkspacedata
@@ -7,7 +7,4 @@
    <FileRef
       location = "group:Demo-macOS/Demo-macOS.xcodeproj">
    </FileRef>
-   <FileRef
-      location = "group:../PocketSVG.xcodeproj">
-   </FileRef>
 </Workspace>
