+++ PocketSVG.xcframework/ios-armv7_arm64/PocketSVG.framework/Headers/PocketSVG.h
@@ -1,0 +1,12 @@
+/*
+ * This file is part of the PocketSVG package.
+ * Copyright (c) Ponderwell, Ariel Elkin, Fjölnir Ásgeirsson, and Contributors
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+#import "SVGBezierPath.h"
+#import "SVGImageView.h"
+#import "SVGLayer.h"
+#import "SVGEngine.h"
