+++ PocketSVG.xcframework/ios-armv7_arm64/PocketSVG.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module PocketSVG {
+  umbrella header "PocketSVG.h"
+
+  export *
+  module * { export * }
+}
