+++ PocketSVG.xcframework/macos-x86_64/PocketSVG.framework/Headers
@@ -1,0 +1,1 @@
+Versions/Current/Headers
