+++ PocketSVG.xcframework/macos-x86_64/PocketSVG.framework/Modules
@@ -1,0 +1,1 @@
+Versions/Current/Modules
