+++ PocketSVG.xcframework/macos-x86_64/PocketSVG.framework/Resources
@@ -1,0 +1,1 @@
+Versions/Current/Resources
