+++ PocketSVG.xcframework/macos-x86_64/PocketSVG.framework/Versions/A/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module PocketSVG {
+  umbrella header "PocketSVG.h"
+
+  export *
+  module * { export * }
+}
