+++ PocketSVG.xcframework/macos-x86_64/PocketSVG.framework/Versions/A/_CodeSignature/CodeResources
@@ -1,0 +1,209 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Resources/Info.plist</key>
+		<data>
+		1iP+CWBv2w/ezWs4V06xjTa86Ic=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/PocketSVG.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			HRPYj7dmj3iW6yrSfA0ZBzqCUks=
+			</data>
+			<key>hash2</key>
+			<data>
+			X70rrWdQd0F3bH5T46uSATXGTQwOsj7TjUQHSMrR6I8=
+			</data>
+		</dict>
+		<key>Headers/SVGBezierPath.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			Sc2JpZfYVLUO7P7Xi+3IdW+c4Gc=
+			</data>
+			<key>hash2</key>
+			<data>
+			5X69hddAJSee1UnGILOXx2mwVG/QOEWBMaNurzB4Bik=
+			</data>
+		</dict>
+		<key>Headers/SVGEngine.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			BP2gW9xpJPIDhfH3MhFlN8BKMDM=
+			</data>
+			<key>hash2</key>
+			<data>
+			0UCtVn0+L6dRHh/UW/oYS45wADHdGYZdRCu9lEjZY4Q=
+			</data>
+		</dict>
+		<key>Headers/SVGImageView.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			JmdUOKOUYJ96PpQgjQfIftrqYOA=
+			</data>
+			<key>hash2</key>
+			<data>
+			wtdUBhPYjgoMaLQYDULPtGa5+QtUuWIeBgqdNr6qRmU=
+			</data>
+		</dict>
+		<key>Headers/SVGLayer.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			0c6YNBodhSRIcg0bP0M0Qq7MpHg=
+			</data>
+			<key>hash2</key>
+			<data>
+			+NA64cLjj7lqq2v42L4EGPgw8aeihBU6uUIqqxSnwjw=
+			</data>
+		</dict>
+		<key>Headers/SVGPortability.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			M6W9nkdT7cgTDS002BSkRO7xubM=
+			</data>
+			<key>hash2</key>
+			<data>
+			uzH3n21AWCpWWB7sz8gYUGEkuE9l3cvBYCwHIfd0Pog=
+			</data>
+		</dict>
+		<key>Modules/module.modulemap</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			Ys47IfP9GvR1Outl8+o6Af1fZzE=
+			</data>
+			<key>hash2</key>
+			<data>
+			cvasCtFXLjpE9j5Nlw8LMTYP87LvkoqfTRLY5dbk3XU=
+			</data>
+		</dict>
+		<key>Resources/Info.plist</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			1iP+CWBv2w/ezWs4V06xjTa86Ic=
+			</data>
+			<key>hash2</key>
+			<data>
+			tVldbp+L1VCSTnu+9UNoQbzp0bJo8A9RFGY5HI/TprE=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^Resources/</key>
+		<true/>
+		<key>^Resources/.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^Resources/.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Resources/Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^(Frameworks|SharedFrameworks|PlugIns|Plug-ins|XPCServices|Helpers|MacOS|Library/(Automator|Spotlight|LoginItems))/</key>
+		<dict>
+			<key>nested</key>
+			<true/>
+			<key>weight</key>
+			<real>10</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^Resources/</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^Resources/.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^Resources/.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Resources/Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^[^/]+$</key>
+		<dict>
+			<key>nested</key>
+			<true/>
+			<key>weight</key>
+			<real>10</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
