+++ PocketSVG.xcframework/macos-x86_64/PocketSVG.framework/Versions/Current
@@ -1,0 +1,1 @@
+A
